backend/pyproject.toml
[tool.poetry]
name = "iaqdb-api"
version = "0.1.0"
description = ""
authors = ["ymarcon <yannick.marcon@epfl.ch>"]
readme = "README.md"
packages = [{include = "iaqdb_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
faker = "^26.0.0"
rasterio = "^1.3.10"
pyproj = "^3.6.1"
requests = "^2.32.3"
sqlmodel = "^0.0.21"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.4.0"
dynaconf = "^3.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
